Add integration tests
Adds integration tests that compare renderer output to known SHA256
hashes. Tests are part of the build system and can be run with ctest.
The test runner is a short python script that provides a cross platform
way to run the render frontend and SHA256 implementation.

Initial hashes were gathered from the renderer branch since it has the
fewest differences from upstream and is more likely to be correct.