main_frontend: rewrite ringbuffer
This commit decouples ringbuffer storage from the type used to access
it. Storage is now an untyped, 64-byte aligned array of bytes.
Ringbuffers are now views into untyped storage and do not necessarily
need to be associated with audio data. Oversampling is no longer a
property of ringbuffers and has been moved back into the frontend.

This commit also enables S32 output for the main frontend.