main_frontend: fix hang on exit on Windows
On Windows we install a Ctrl-C handler to make sure that the event loop
always receives an SDL_QUIT event. This is what normally happens on
other platforms but only some Windows environments (for instance, a
mingw64 shell). If the program is run from cmd or explorer, SDL_QUIT is
never sent and the program hangs.