main_frontend: ensure program is always responsive
This commit removes the MCU-wide lock that was shared between the MCU
thread and the UI thread. Instead, the lock has been moved into the LCD
where the MCU can take it for very brief periods only when updating LCD
state.

The UI thread will now drop a frame only if the MCU is currently
updating LCD state. Otherwise, it will take the lock, quickly copy the
state necessary for rendering a frame, and release the lock. This also
means that the LCD display will never freeze for extended periods of
time.

Regardless of whether or not the LCD is enabled, the UI thread (which
also handles system-wide events) will never be blocked for long and will
always respond to quit messages from the user.