cmake: change project name to nuked-sc55
The exact spelling of the name is significant for controlling install
directories. Since the standard frontend looks for data in
`../share/nuked-sc55`, `nuked-sc55` seems like the most reasonable
choice.

In a future commit, documentation will be installed to
`CMAKE_INSTALL_DOCDIR`. This variable uses the project name in the path,
and we do not want to create both `../share/nuked-sc55` and
`../share/doc/NukedSC55`.